kamheat: support 16bit status fields
We just bought a bunch of Kamstrup Multical 303, the type info as
printed on the device is 303W40276. The supplier mounted one of them in
a reversed direction, which is shown on the display as info code
0004000. It appears that this status field is reported as DIF `0x02`,
VIF `0xFF`, VIFE `0x22`, but the `kamheat` driver currently assumes that
it's sent as a 32bit `difvifkey = 04FF22` (or as `ErrorFlags), and these
are not being sent by our devices.

We have some docs available, they list some bit flag names. Those which
are listed in our docs have the same meaning, but some fields which are
defined in the existing driver are not listed in our docs. The good
thing is that simply extending the `match` field apparently does the
job.

Co-authored-by: Fredrik Öhrström <oehrstroem@gmail.com>
Bug: https://github.com/wmbusmeters/wmbusmeters/issues/1653
Link: https://www.sge.cz/images/MC303/MULTICAL303-Technical-Description-English.pdf